build: update nightly and dev flatpak.
In particular a bunch of dependencies were moved to the BaseApp.
Dev flatpak is actually a bit useless now (since last dev release is
older than stable). But I keep the dev manifest around to make it easier
to update for the future when we'll have a new dev release.


Former-commit-id: 1ef0f50698132e63d182a147ad538371e01b0b12