app: connect GimpHistogramView to bg-histogram "notify" signal
In GimpHistogramView, connect the view to the bg-histogram's
"notify" signal, and update the view in response.  Previously, we
only updated the view in response to the main histogram's "notify"
signal.

The curves tool only uses the bg-histogram, and after the switch to
asynchronous histogram calculation, in commit
db827608b410a8f9d6157108ae5efb041214d824 [formerly 49382e53d55d3be65135130a2898038aad1a3719], it failed to update the
histogram view after the calculation was complete.


Former-commit-id: 4ad72cfc77f870d801c4a0ab49958f8fb2b057b7