desktop: remove the double <releases> trick.
This was added in commits ba6f0cf54a [formerly 447d9bbc56] and 3700b0d6e3 [formerly fb6328b9ad].
With it, the validation succeeds. Still I am actually not sure that is
proper appdata format. Also we have issues of metadata not shown for our
flatpak package, and I am wondering if this could be the issue.
So let's get rid of this hack.

Also temporarily commenting out the appdata unit test because it fails
without, because of a bug in appstream-util. This bug has been fixed
since, but there are no releases containing this fix yet.
See: https://github.com/hughsie/appstream-glib/issues/234
We'll reactivate the tests later.

Former-commit-id: 0492157dd2851a6bfa42af106d5d3809b9924fd2