build: add my cron script to build dev flatpak regularly.
This is the script I have been personally using for some time now to
build the devel and nightly flatpak on my server regularly through cron
files.
It is not the best script, arguably, and we can most likely do better.
But for now, it worked for me and is a first step toward maybe making
official nightly flatpak builds soon.


Former-commit-id: b748dea801dd0c9982ebd52450e3534850f0e132