build: add a flatpak for the gtk3-port branch.
This flatpak is tested and working.
I have not yet verified though if it has the maximum options enabled
(guessing not), but only that we had at least the required dependencies
at least.

When we will merge, it will become the nightly flatpak and the current
nightly can be used to follow the 2.10 (future) branch.


Former-commit-id: 1d30426102d5559401b8f4eb3c39e254eaf99f02