app: log actions upon activation before emitting "selected" signal
In GimpAction, instead of connecting the action-history log
function to the action's "activate" signal as a user-provided
handler, call it directly from the default handler.

In subclasses of GimpAction, chain to the parent's activate()
function before emitting the "selected" signal, so that we always
log the action in the history before responding to it.

This allows us to avoid the hack in commit
668b3d08e63787916b40b83877f6a2d8f293bf54 [formerly 6544ce4301fa43636b4c282c7b4815e18fd653d1].


Former-commit-id: 114d49510f2031e00aafafb6c4666b5213d0d0cf